I moved some fuction anoder file
